chore: 과제 상세 페이지 라우팅
